requirements.txt
altgraph==0.17.3
certifi==2023.7.22
charset-normalizer==3.2.0
idna==3.4
macholib==1.16.2
pefile==2023.2.7
protobuf==4.24.2
python-dotenv==1.0.0
pywin32-ctypes==0.2.2
PyYAML==6.0.1
requests==2.31.0
urllib3==2.0.4
websocket-client==1.6.2
